Update deployment targets to iOS 13.  Rename iOS13NativeSwiftTemplate to iOSNativeSwiftTemplate.
