+++ MobileSyncExplorerReactNative/package.json
@@ -15,7 +15,7 @@
         "@react-navigation/native": "^5.2.1",
         "@react-navigation/stack": "^5.2.16",
         "react": "16.13.1",
-        "react-native": "0.63.2",
+        "react-native": "0.63.3",
         "react-native-force": "git+https://github.com/forcedotcom/SalesforceMobileSDK-ReactNative.git#dev",
         "typescript": "^3.9.3",
         "react-native-elements": "^0.19.1",
