+++ ReactNativeTemplate/package.json
@@ -15,7 +15,7 @@
         "@react-navigation/native": "^5.2.0",
         "@react-navigation/stack": "^5.2.15",
         "react": "16.13.1",
-        "react-native": "0.63.2",
+        "react-native": "0.63.3",
         "react-native-force": "git+https://github.com/forcedotcom/SalesforceMobileSDK-ReactNative.git#dev",
         "typescript": "^3.9.3",
         "react-native-gesture-handler": "^1.6.1",
