--- iOS13NativeSwiftTemplate/.gitignore
@@ -1,26 +1,0 @@
-DerivedData
-project.xcworkspace
-*.iml
-*.xcworkspace/
-*~
-.DS_Store
-.github/
-.gradle/
-.idea/
-.metadata/
-.settings/
-Podfile.lock
-Pods/
-bin/
-bower_components/
-build/
-coverage.em
-coverage/
-instrumentation/
-junit-reports/
-local.properties
-node_modules/
-mobile_sdk/
-proguard-project.txt
-xcuserdata
-package-lock.json
