--- iOS13NativeSwiftTemplate/Podfile
@@ -1,28 +1,0 @@
-platform :ios, '13.0'
-
-project 'iOS13NativeSwiftTemplate.xcodeproj'
-target 'iOS13NativeSwiftTemplate' do
-
-source 'https://cdn.cocoapods.org/'
-
-use_frameworks!
-
-pod 'SalesforceSDKCommon', :path => 'mobile_sdk/SalesforceMobileSDK-iOS'
-pod 'SalesforceAnalytics', :path => 'mobile_sdk/SalesforceMobileSDK-iOS'
-pod 'SalesforceSDKCore', :path => 'mobile_sdk/SalesforceMobileSDK-iOS'
-pod 'SmartStore', :path => 'mobile_sdk/SalesforceMobileSDK-iOS'
-pod 'MobileSync', :path => 'mobile_sdk/SalesforceMobileSDK-iOS'
-
-end
-
-# Comment the following if you do not want the SDK to emit signpost events for instrumentation. Signposts are  enabled for non release version of the app.
- post_install do |installer_representation|
-       installer_representation.pods_project.targets.each do |target|
-           target.build_configurations.each do |config|
-               if config.name == 'Debug'
-                   config.build_settings['GCC_PREPROCESSOR_DEFINITIONS'] ||= ['$(inherited)', 'DEBUG=1','SIGNPOST_ENABLED=1']
-                   config.build_settings['OTHER_SWIFT_FLAGS'] = ['$(inherited)', '-DDEBUG','-DSIGNPOST_ENABLED']
-               end
-           end
-       end
-  end
