--- iOS13NativeSwiftTemplate/iOS13NativeSwiftTemplate/bootconfig.plist
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>remoteAccessConsumerKey</key>
-	<string>3MVG9Iu66FKeHhINkB1l7xt7kR8czFcCTUhgoA8Ol2Ltf1eYHOU4SqQRSEitYFDUpqRWcoQ2.dBv_a1Dyu5xa</string>
-	<key>oauthRedirectURI</key>
-	<string>testsfdc:///mobilesdk/detect/oauth/done</string>
-	<key>oauthScopes</key>
-	<array>
-		<string>web</string>
-		<string>api</string>
-	</array>
-	<key>shouldAuthenticate</key>
-	<true/>
-</dict>
-</plist>
