--- iOS13NativeSwiftTemplate/iOS13NativeSwiftTemplate/iOS13NativeSwiftTemplate.entitlements
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>keychain-access-groups</key>
-	<array>
-		<string>$(AppIdentifierPrefix)com.salesforce.ios13nativeswifttemplate.iOS13NativeSwiftTemplate</string>
-	</array>
-</dict>
-</plist>
