--- iOS13NativeSwiftTemplate/package.json
@@ -1,8 +1,0 @@
-{
-    "name": "iOS13NativeSwiftTemplate",
-    "version": "0.0.1",
-    "private": true,
-    "sdkDependencies": {
-        "SalesforceMobileSDK-iOS": "https://github.com/forcedotcom/SalesforceMobileSDK-iOS.git#dev"
-    }
-}
