+++ iOSNativeSwiftTemplate/iOSNativeSwiftTemplate/ContactDetailModel.swift
