+++ iOSNativeSwiftTemplate/iOSNativeSwiftTemplate/InitialViewController.swift
@@ -1,5 +1,5 @@
 /*
-Copyright (c) 2014-present, salesforce.com, inc. All rights reserved.
+Copyright (c) 2019-present, salesforce.com, inc. All rights reserved.
 
 Redistribution and use of this software in source and binary forms, with or without modification,
 are permitted provided that the following conditions are met:
@@ -25,5 +25,5 @@ WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH 
 import Foundation
 import UIKit
 
-class InitialViewController: UIViewController {
+class InitialViewController : UIViewController {
 }
