+++ iOSNativeSwiftTemplate/iOSNativeSwiftTemplate/main.swift
@@ -1,5 +1,5 @@
 /*
-Copyright (c) 2015-present, salesforce.com, inc. All rights reserved.
+Copyright (c) 2019-present, salesforce.com, inc. All rights reserved.
 
 Redistribution and use of this software in source and binary forms, with or without modification,
 are permitted provided that the following conditions are met:
