+++ setversion.sh
@@ -66,7 +66,6 @@ update_package_json "./MobileSyncExplorerReactNative/package.json"  "${SDK_TAG}"
 update_package_json "./MobileSyncExplorerSwift/package.json"  "${SDK_TAG}"
 update_package_json "./ReactNativeTemplate/package.json"  "${SDK_TAG}"
 update_package_json "./ReactNativeDeferredTemplate/package.json"  "${SDK_TAG}"
-update_package_json "./iOS13NativeSwiftTemplate/package.json"  "${SDK_TAG}"
 update_package_json "./iOSIDPTemplate/package.json"  "${SDK_TAG}"
 update_package_json "./iOSNativeSwiftEncryptedNotificationTemplate/package.json"  "${SDK_TAG}"
 update_package_json "./iOSNativeSwiftTemplate/package.json"  "${SDK_TAG}"
